feat: Rename logging module to avoid shadowing standard library and update imports
